Moving getApiContext method to functions
